pyproject.toml
[tool.poetry]
name = "ai-agent-hackaton"
version = "0.1.0"
description = ""
authors = ["Ayobami Alaran A. <ayobamidele006@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
sparky-utils = "^1.0.9"
django-cors-headers = "^4.4.0"
python-dotenv = "^1.0.1"
azure-ai-inference = "^1.0.0b4"
composio-openai = "^0.5.19"
pytz = "^2024.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
